backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-auth-ldap==4.6.0
django-cors-headers==4.3.1
django-celery-beat==2.6.0
dj-database-url==2.1.0
celery[redis]==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
ldap3==2.9.1
Pillow==10.2.0
qrcode==7.4.2
python-dotenv==1.0.1
gunicorn==21.2.0
cryptography==42.0.8
